Fixed char.c msvc compiler warnings
Thanks to LuLu for bringing it up.

+++ src/char/char.c
@@ -1897,7 +1897,7 @@ int mmo_char_tobuf(uint8* buffer, struct mmo_charstatus* p) {
 	offset += MAP_NAME_LENGTH_EXT;
 #endif
 #if PACKETVER >= 20100803
-	WBUFL(buf,124) = p->delete_date;
+	WBUFL(buf,124) = (int)p->delete_date;
 	offset += 4;
 #endif
 #if PACKETVER >= 20110111
@@ -3566,9 +3566,9 @@ void char_delete2_ack(int fd, int char_id, uint32 result, time_t delete_date)
 	WFIFOL(fd,2) = char_id;
 	WFIFOL(fd,6) = result;
 #if PACKETVER >= 20130000
-	WFIFOL(fd,10) = delete_date - time(NULL);
+	WFIFOL(fd,10) = (int)(delete_date - time(NULL));
 #else
-	WFIFOL(fd,10) = delete_date;
+	WFIFOL(fd,10) = (int)delete_date;
 
 #endif
 	WFIFOSET(fd,14);
